Moved Mir2Llvm unit tests to their own file to make them easier to use
